StallionAnalytics/requirements.txt
streamlit
pandas
plotly
streamlit-extras
pydantic
openai
watchdog
tabulate
google-generativeai